add new hook useSmartPlaylists - extracted logic from useTagTada
